Split ofp conversion kernel
