+++ src/cvt/README.md
@@ -1,4 +1,4 @@
-# OFP (*Open Compute Project* Floating Point) Conversion Kernels
+#OFP(*Open Compute Project *Floating Point) Conversion Kernels
 
 The OFP conversion kernels provide vectorized conversion functions between various floating-point formats and 8-bit/4-bit OFP formats using RISC-V Zvfofp8min and Zvfofp4min extensions. These kernels enable efficient data type conversions for neural network workloads that utilize low-precision floating-point formats.
 
@@ -14,7 +14,7 @@ Narrowing conversions reduce precision and dynamic range from standard formats t
 
 **Basic conversion (F32 to E4M3):**
 ```c
-/* in cvt_zvfofp8min.h */
+/* in cvt_f32_f8_zvfofp8min.h */
 
 /** F32 to E4M3 conversion with scaling.
  *
@@ -69,7 +69,7 @@ The OFP4 conversion functions support widening conversions from 4-bit E2M1 forma
 
 **E2M1 to E4M3 conversion:**
 ```c
-/* in cvt_zvfofp4min.h */
+/* in cvt_f4_f8_zvfofp4min.h */
 
 /** E2M1 to E4M3 conversion.
  *
